+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 0.1.1
 
 * Update dependencies
+* `Triggers`
+  * `SelectorWithTimer`
+    * Opportunity to get schedule of posts using `publishing_autoschedule` command
 
 ## 0.0.10